Move footer components to single directory and remove unneeded stories
